Make a separate package
The separate package includes now only what's needed for scfinder.
Hoping nothing is missing, more commits might follow.

eewamps will be moved to main seiscomp in the near futur.

The repo becomes now sed-addons and it is the right place for
seiscomp-based SED products with restricted distribution (in
dev or private).