Initial contribution for sceewlog
scvsmaglog will be replaced by sceewlog. This will be developed
while anticipating further developments for scvsmag and the EEW
decision module and this is recycling a lot of the existing
scvsmag codes.

In the meantime, changed will be made in the eew/scvsmag-dev
directory so that the existing scvsmag and the new modules can all be
compiled together and compared.

Further info are in apps/eew/scvsmag-dev/Specifications.md