Fix pref Or/Ma issue
sceelog send each magnitude update as an xml message via activeMQ.
But it did not check if the magnitude is preferred.
Then, in the case when several locators are active, and when the
preferred origin and magnitude is quickly changing, the preferred
magnitude can end up being missing from the xml message, or different
from the magnitude update.

This commit ensures that the origin and magnitude objects of the
magnitude update :
- are part of the xml message
- are setup as preferred.